refactor: move getProjects logic to utils
