+++ src/components/Button/button.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '../../../jest-configs/test-utils';
 import { theme } from '../../assets/styles/theme';
 
-import Button from '../Button/Button';
+import Button from './Button';
 
 describe('Button ', () => {
     it('correctly renders button to the page', () => {
