+++ src/data/projects/portfolio.json
@@ -13,6 +13,5 @@
     ],
     "codeLink": "https://github.com/Dabgan/Portfolio",
     "demoLink": "https://www.gabrieldaniluk.com/",
-    "coverImg": "../../assets/images/portfolio.png",
-    "inverted": true
+    "coverImg": "../../assets/images/portfolio.png"
 }
