--- src/data/projects/quiz.json
@@ -1,11 +1,0 @@
-{
-    "order": 3,
-    "title": "Quiz App",
-    "introduction": "This app allows you to solve quizzes from selected categories.",
-    "description": "This app allows you to solve quizzes from selected categories. When I was creating this project I was just starting with React, so the entire application is written with Class Components. I used SCSS for styling.",
-    "technologies": ["React", "SCSS", "Github-pages"],
-    "codeLink": "https://github.com/Dabgan/Project06-React_Quiz",
-    "demoLink": "https://dabgan.github.io/Project06-React_Quiz/",
-    "coverImg": "../../assets/images/quiz.png",
-    "inverted": false
-}
