+++ src/data/projects/snowshop.json
@@ -2,7 +2,7 @@
     "order": 2,
     "title": "Snowshop",
     "introduction": "My biggest project so far. A mockup of a fully functioning online shop.",
-    "description": "A mockup of a fully functioning online store using a Firebase-based backend. The application includes such functionalities as products filtering, a shopping cart or an order form. For styling, I used styled-components. While creating this project, I learned how to solve advanced React problems.",
+    "description": "A mockup of a fully functioning online store using a Firebase-based backend. The application includes such functionalities as products filtering, a shopping cart or an order form. For styling, I used styled-components.",
     "technologies": [
         "React & React Hooks",
         "Styled Components",
@@ -13,6 +13,5 @@
     ],
     "codeLink": "https://github.com/Dabgan/SnowShop",
     "demoLink": "https://snow-shop.netlify.app/",
-    "coverImg": "../../assets/images/snowshop.png",
-    "inverted": true
+    "coverImg": "../../assets/images/snowshop.png"
 }
