+++ src/data/projects/timeline.json
@@ -11,6 +11,5 @@
     ],
     "codeLink": "https://github.com/Dabgan/github-timeline",
     "demoLink": "https://commits-timeline.netlify.app/",
-    "coverImg": "../../assets/images/timeline.png",
-    "inverted": false
+    "coverImg": "../../assets/images/timeline.png"
 }
